+++ mods/BrewLAN_Plenae/Logger/mod_info.lua
@@ -1,6 +1,6 @@
 name = "Debug Tools"
-uid = "2529ea71-93ef-41a6-b552-LOGS0000000008"
-version = 0.8
+uid = "2529ea71-93ef-41a6-b552-LOGS0000000009"
+version = 0.9
 copyright = "© 2017–2018 Sean Wheeldon"
 description = "Debug and testing tools. Not for real matches."
 author = "Balthazar"