+++ mods/BrewLAN_Units/BrewTurrets/units/SEB2402/SEB2402_script.lua
@@ -1,0 +1,11 @@
+local TStructureUnit = import('/lua/terranunits.lua').TStructureUnit
+local MaelstromDeathLaser = import('/lua/sim/defaultweapons.lua').MaelstromDeathLaser
+
+SEB2402 = Class(TStructureUnit) {
+    Weapons = {
+        MainGun = Class(MaelstromDeathLaser){
+            --FxMuzzleFlash = {},
+        },
+    },
+}
+TypeClass = SEB2402